requirements.txt
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
boto3==1.34.47
botocore==1.34.47
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.2
cssselect==1.2.0
exceptiongroup==1.2.0
filelock==3.13.1
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==5.1.0
outcome==1.3.0.post0
packaging==23.2
parsel==1.8.1
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
s3transfer==0.10.0
Scrapy==2.11.0
scrapy-selenium==0.0.7
scrapy-zyte-smartproxy==2.3.2
selenium==4.17.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tldextract==5.1.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.0.7
w3lib==2.1.2
webdriver-manager==4.0.1
wsproto==1.2.0
zope.interface==6.1
openai==1.30.4